+++ js/scripts.js
@@ -1,7 +1,8 @@
+// import './_scroll.js';
 import './_svgs.js';
 
-// import im from '../node_modules/include-media-export/include-media.js';
-import fitvids from '../node_modules/fitvids/dist/fitvids.js';
+// import im from 'include-media-export/include-media.js';
+import fitvids from 'fitvids/dist/fitvids.js';
 
 (function() {
 	// Remove the 'no-js' <body> class